requirements.txt
streamlit
yfinance
sentence-transformers
faiss-cpu
openai
transformers
torch
numpy
